requirements.txt
ansicolors==1.1.8
ansiwrap==0.8.4
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.10.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.7
django-classy-tags==4.1.0
django-extensions==4.1
django-js-asset==3.1.2
django-mptt==0.16.0
django-nyt==1.4.2
django-sekizai==4.1.0
djangorestframework==3.16.1
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.15
lark==1.3.0
Markdown==3.9
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
-e git+https://github.com/ghpz5649xx-blip/notelib.git@75d869c577a434525ba00c184dad93ccfe5a34e3#egg=notelib_core
numpy==2.3.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
papermill==2.6.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
pymdown-extensions==10.5
pyrsistent==0.20.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
sorl-thumbnail==12.11.0
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.1.2
terminado==0.18.1
textwrap3==0.9.2
tinycss2==1.4.0
tomli==2.3.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.14
wiki==0.12.2
zipp==3.23.0
zstandard==0.25.0
